Crate input_sys

Structs

__va_list_tag

libinput

libinput_device

libinput_device_group

libinput_event

libinput_event_device_notify

libinput_event_gesture

libinput_event_keyboard

libinput_event_pointer

libinput_event_switch

libinput_event_tablet_pad

libinput_event_tablet_tool

libinput_event_touch

libinput_interface
@ingroup base @struct libinput_interface

libinput_seat

libinput_tablet_pad_mode_group

libinput_tablet_tool

udev

udev_device

Enums

libinput_button_state
@ingroup device

libinput_config_accel_profile
@ingroup config

libinput_config_click_method
@ingroup config

libinput_config_drag_lock_state
@ingroup config

libinput_config_drag_state
@ingroup config

libinput_config_dwt_state
@ingroup config

libinput_config_middle_emulation_state
@ingroup config

libinput_config_scroll_method
@ingroup config

libinput_config_status
@ingroup config

libinput_config_tap_button_map
@ingroup config

libinput_config_tap_state
@ingroup config

libinput_device_capability
@ingroup device

libinput_event_type
@ingroup base

libinput_key_state
@ingroup device

libinput_log_priority
@ingroup base

libinput_pointer_axis
@ingroup device

libinput_pointer_axis_source
@ingroup device

libinput_switch
@ingroup device

libinput_switch_state
@ingroup device

libinput_tablet_pad_ring_axis_source
@ingroup event_tablet_pad

libinput_tablet_pad_strip_axis_source
@ingroup event_tablet_pad

libinput_tablet_tool_proximity_state
@ingroup device

libinput_tablet_tool_tip_state
@ingroup device

libinput_tablet_tool_type
@ingroup device

Constants

libinput_config_send_events_mode_LIBINPUT_CONFIG_SEND_EVENTS_DISABLED

libinput_config_send_events_mode_LIBINPUT_CONFIG_SEND_EVENTS_DISABLED_ON_EXTERNAL_MOUSE

libinput_config_send_events_mode_LIBINPUT_CONFIG_SEND_EVENTS_ENABLED

libinput_led_LIBINPUT_LED_CAPS_LOCK

libinput_led_LIBINPUT_LED_NUM_LOCK

libinput_led_LIBINPUT_LED_SCROLL_LOCK

Functions

libinput_config_status_to_str⚠
@ingroup config

libinput_device_config_accel_get_default_profile⚠
@ingroup config

libinput_device_config_accel_get_default_speed⚠
@ingroup config

libinput_device_config_accel_get_profile⚠
@ingroup config

libinput_device_config_accel_get_profiles⚠
@ingroup config

libinput_device_config_accel_get_speed⚠
@ingroup config

libinput_device_config_accel_is_available⚠
@ingroup config

libinput_device_config_accel_set_profile⚠
@ingroup config

libinput_device_config_accel_set_speed⚠
@ingroup config

libinput_device_config_calibration_get_default_matrix⚠
@ingroup config

libinput_device_config_calibration_get_matrix⚠
@ingroup config

libinput_device_config_calibration_has_matrix⚠
@ingroup config

libinput_device_config_calibration_set_matrix⚠
@ingroup config

libinput_device_config_click_get_default_method⚠
@ingroup config

libinput_device_config_click_get_method⚠
@ingroup config

libinput_device_config_click_get_methods⚠
@ingroup config

libinput_device_config_click_set_method⚠
@ingroup config

libinput_device_config_dwt_get_default_enabled⚠
@ingroup config

libinput_device_config_dwt_get_enabled⚠
@ingroup config

libinput_device_config_dwt_is_available⚠
@ingroup config

libinput_device_config_dwt_set_enabled⚠
@ingroup config

libinput_device_config_left_handed_get⚠
@ingroup config

libinput_device_config_left_handed_get_default⚠
@ingroup config

libinput_device_config_left_handed_is_available⚠
@ingroup config

libinput_device_config_left_handed_set⚠
@ingroup config

libinput_device_config_middle_emulation_get_default_enabled⚠
@ingroup config

libinput_device_config_middle_emulation_get_enabled⚠
@ingroup config

libinput_device_config_middle_emulation_is_available⚠
@ingroup config

libinput_device_config_middle_emulation_set_enabled⚠
@ingroup config

libinput_device_config_rotation_get_angle⚠
@ingroup config

libinput_device_config_rotation_get_default_angle⚠
@ingroup config

libinput_device_config_rotation_is_available⚠
@ingroup config

libinput_device_config_rotation_set_angle⚠
@ingroup config

libinput_device_config_scroll_get_button⚠
@ingroup config

libinput_device_config_scroll_get_default_button⚠
@ingroup config

libinput_device_config_scroll_get_default_method⚠
@ingroup config

libinput_device_config_scroll_get_default_natural_scroll_enabled⚠
@ingroup config

libinput_device_config_scroll_get_method⚠
@ingroup config

libinput_device_config_scroll_get_methods⚠
@ingroup config

libinput_device_config_scroll_get_natural_scroll_enabled⚠
@ingroup config

libinput_device_config_scroll_has_natural_scroll⚠
@ingroup config

libinput_device_config_scroll_set_button⚠
@ingroup config

libinput_device_config_scroll_set_method⚠
@ingroup config

libinput_device_config_scroll_set_natural_scroll_enabled⚠
@ingroup config

libinput_device_config_send_events_get_default_mode⚠
@ingroup config

libinput_device_config_send_events_get_mode⚠
@ingroup config

libinput_device_config_send_events_get_modes⚠
@ingroup config

libinput_device_config_send_events_set_mode⚠
@ingroup config

libinput_device_config_tap_get_button_map⚠
@ingroup config

libinput_device_config_tap_get_default_button_map⚠
@ingroup config

libinput_device_config_tap_get_default_drag_enabled⚠
@ingroup config

libinput_device_config_tap_get_default_drag_lock_enabled⚠
@ingroup config

libinput_device_config_tap_get_default_enabled⚠
@ingroup config

libinput_device_config_tap_get_drag_enabled⚠
@ingroup config

libinput_device_config_tap_get_drag_lock_enabled⚠
@ingroup config

libinput_device_config_tap_get_enabled⚠
@ingroup config

libinput_device_config_tap_get_finger_count⚠
@ingroup config

libinput_device_config_tap_set_button_map⚠
@ingroup config

libinput_device_config_tap_set_drag_enabled⚠
@ingroup config

libinput_device_config_tap_set_drag_lock_enabled⚠
@ingroup config

libinput_device_config_tap_set_enabled⚠
@ingroup config

libinput_device_get_context⚠
@ingroup device

libinput_device_get_device_group⚠
@ingroup device

libinput_device_get_id_product⚠
@ingroup device

libinput_device_get_id_vendor⚠
@ingroup device

libinput_device_get_name⚠
@ingroup device

libinput_device_get_output_name⚠
@ingroup device

libinput_device_get_seat⚠
@ingroup device

libinput_device_get_size⚠
@ingroup device

libinput_device_get_sysname⚠
@ingroup device

libinput_device_get_udev_device⚠
@ingroup device

libinput_device_get_user_data⚠
@ingroup device

libinput_device_group_get_user_data⚠
@ingroup device

libinput_device_group_ref⚠
@ingroup device

libinput_device_group_set_user_data⚠
@ingroup device

libinput_device_group_unref⚠
@ingroup device

libinput_device_has_capability⚠
@ingroup device

libinput_device_keyboard_has_key⚠
@ingroup device

libinput_device_led_update⚠
@ingroup device

libinput_device_pointer_has_button⚠
@ingroup device

libinput_device_ref⚠
@ingroup device

libinput_device_set_seat_logical_name⚠
@ingroup device

libinput_device_set_user_data⚠
@ingroup device

libinput_device_tablet_pad_get_mode_group⚠
@ingroup tablet_pad_modes

libinput_device_tablet_pad_get_num_buttons⚠
@ingroup device

libinput_device_tablet_pad_get_num_mode_groups⚠
@ingroup tablet_pad_modes

libinput_device_tablet_pad_get_num_rings⚠
@ingroup device

libinput_device_tablet_pad_get_num_strips⚠
@ingroup device

libinput_device_unref⚠
@ingroup device

libinput_dispatch⚠
@ingroup base

libinput_event_destroy⚠
@ingroup event

libinput_event_device_notify_get_base_event⚠
@ingroup event

libinput_event_gesture_get_angle_delta⚠
@ingroup event_gesture

libinput_event_gesture_get_base_event⚠
@ingroup event_gesture

libinput_event_gesture_get_cancelled⚠
@ingroup event_gesture

libinput_event_gesture_get_dx⚠
@ingroup event_gesture

libinput_event_gesture_get_dx_unaccelerated⚠
@ingroup event_gesture

libinput_event_gesture_get_dy⚠
@ingroup event_gesture

libinput_event_gesture_get_dy_unaccelerated⚠
@ingroup event_gesture

libinput_event_gesture_get_finger_count⚠
@ingroup event_gesture

libinput_event_gesture_get_scale⚠
@ingroup event_gesture

libinput_event_gesture_get_time⚠
@ingroup event_gesture

libinput_event_gesture_get_time_usec⚠
@ingroup event_gesture

libinput_event_get_context⚠
@ingroup event

libinput_event_get_device⚠
@ingroup event

libinput_event_get_device_notify_event⚠
@ingroup event

libinput_event_get_gesture_event⚠
@ingroup event

libinput_event_get_keyboard_event⚠
@ingroup event

libinput_event_get_pointer_event⚠
@ingroup event

libinput_event_get_switch_event⚠
@ingroup event

libinput_event_get_tablet_pad_event⚠
@ingroup event

libinput_event_get_tablet_tool_event⚠
@ingroup event

libinput_event_get_touch_event⚠
@ingroup event

libinput_event_get_type⚠
@ingroup event

libinput_event_keyboard_get_base_event⚠
@ingroup event_keyboard

libinput_event_keyboard_get_key⚠
@ingroup event_keyboard

libinput_event_keyboard_get_key_state⚠
@ingroup event_keyboard

libinput_event_keyboard_get_seat_key_count⚠
@ingroup event_keyboard

libinput_event_keyboard_get_time⚠
@ingroup event_keyboard

libinput_event_keyboard_get_time_usec⚠
@ingroup event_keyboard

libinput_event_pointer_get_absolute_x⚠
@ingroup event_pointer

libinput_event_pointer_get_absolute_x_transformed⚠
@ingroup event_pointer

libinput_event_pointer_get_absolute_y⚠
@ingroup event_pointer

libinput_event_pointer_get_absolute_y_transformed⚠
@ingroup event_pointer

libinput_event_pointer_get_axis_source⚠
@ingroup event_pointer

libinput_event_pointer_get_axis_value⚠
@ingroup event_pointer

libinput_event_pointer_get_axis_value_discrete⚠
@ingroup event_pointer

libinput_event_pointer_get_base_event⚠
@ingroup event_pointer

libinput_event_pointer_get_button⚠
@ingroup event_pointer

libinput_event_pointer_get_button_state⚠
@ingroup event_pointer

libinput_event_pointer_get_dx⚠
@ingroup event_pointer

libinput_event_pointer_get_dx_unaccelerated⚠
@ingroup event_pointer

libinput_event_pointer_get_dy⚠
@ingroup event_pointer

libinput_event_pointer_get_dy_unaccelerated⚠
@ingroup event_pointer

libinput_event_pointer_get_seat_button_count⚠
@ingroup event_pointer

libinput_event_pointer_get_time⚠
@ingroup event_pointer

libinput_event_pointer_get_time_usec⚠
@ingroup event_pointer

libinput_event_pointer_has_axis⚠
@ingroup event_pointer

libinput_event_switch_get_base_event⚠
@ingroup event_switch

libinput_event_switch_get_switch⚠
@ingroup event_switch

libinput_event_switch_get_switch_state⚠
@ingroup event_switch

libinput_event_switch_get_time⚠
@ingroup event_switch

libinput_event_switch_get_time_usec⚠
@ingroup event_switch

libinput_event_tablet_pad_get_base_event⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_button_number⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_button_state⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_mode⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_mode_group⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_ring_number⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_ring_position⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_ring_source⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_strip_number⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_strip_position⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_strip_source⚠
@ingroup event_tablet_pad

libinput_event_tablet_pad_get_time⚠
@ingroup event_tablet

libinput_event_tablet_pad_get_time_usec⚠
@ingroup event_tablet_pad

libinput_event_tablet_tool_distance_has_changed⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_base_event⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_button⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_button_state⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_distance⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_dx⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_dy⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_pressure⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_proximity_state⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_rotation⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_seat_button_count⚠

@ingroup event_tablet *

For the button of a @ref LIBINPUT_EVENT_TABLET_TOOL_BUTTON event, return the total

number of buttons pressed on all devices on the associated seat after the

the event was triggered. * " @note It is an application bug to call this function for events other than

@ref LIBINPUT_EVENT_TABLET_TOOL_BUTTON. For other events, this function returns 0. *

@param event The libinput tablet tool event

@return the seat wide pressed button count for the key of this event

libinput_event_tablet_tool_get_slider_position⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_tilt_x⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_tilt_y⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_time⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_time_usec⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_tip_state⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_tool⚠

@ingroup event_tablet *

Returns the tool that was in use during this event. *

The returned tablet tool is not refcounted and may become invalid after

the next call to libinput. Use libinput_tablet_tool_ref() and

libinput_tablet_tool_unref() to continue using the handle outside of the

immediate scope. *

If the caller holds at least one reference, this struct is used

whenever the tools enters proximity again. *

@note Physical tool tracking requires hardware support. If unavailable,

libinput creates one tool per type per tablet. See @ref

tablet-serial-numbers for more details. *

@param event The libinput tablet tool event

@return The new tool triggering this event

libinput_event_tablet_tool_get_wheel_delta⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_wheel_delta_discrete⚠

@ingroup event_tablet *

Return the delta for the wheel in discrete steps (e.g. wheel clicks).

@param event The libinput tablet tool event

@return The delta of the wheel, in discrete steps, compared to the last event *

@see libinput_event_tablet_tool_get_wheel_delta_discrete

libinput_event_tablet_tool_get_x⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_x_transformed⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_y⚠
@ingroup event_tablet

libinput_event_tablet_tool_get_y_transformed⚠
@ingroup event_tablet

libinput_event_tablet_tool_pressure_has_changed⚠
@ingroup event_tablet

libinput_event_tablet_tool_rotation_has_changed⚠
@ingroup event_tablet

libinput_event_tablet_tool_slider_has_changed⚠
@ingroup event_tablet

libinput_event_tablet_tool_tilt_x_has_changed⚠
@ingroup event_tablet

libinput_event_tablet_tool_tilt_y_has_changed⚠
@ingroup event_tablet

libinput_event_tablet_tool_wheel_has_changed⚠
@ingroup event_tablet

libinput_event_tablet_tool_x_has_changed⚠
@ingroup event_tablet

libinput_event_tablet_tool_y_has_changed⚠
@ingroup event_tablet

libinput_event_touch_get_base_event⚠
@ingroup event_touch

libinput_event_touch_get_seat_slot⚠
@ingroup event_touch

libinput_event_touch_get_slot⚠
@ingroup event_touch

libinput_event_touch_get_time⚠
@ingroup event_touch

libinput_event_touch_get_time_usec⚠
@ingroup event_touch

libinput_event_touch_get_x⚠
@ingroup event_touch

libinput_event_touch_get_x_transformed⚠
@ingroup event_touch

libinput_event_touch_get_y⚠
@ingroup event_touch

libinput_event_touch_get_y_transformed⚠
@ingroup event_touch

libinput_get_event⚠
@ingroup base

libinput_get_fd⚠
@ingroup base

libinput_get_user_data⚠
@ingroup base

libinput_log_get_priority⚠
@ingroup base

libinput_log_set_handler⚠
@ingroup base

libinput_log_set_priority⚠
@ingroup base

libinput_next_event_type⚠
@ingroup base

libinput_path_add_device⚠
@ingroup base

libinput_path_create_context⚠
@ingroup base

libinput_path_remove_device⚠
@ingroup base

libinput_ref⚠
@ingroup base

libinput_resume⚠
@ingroup base

libinput_seat_get_context⚠
@ingroup seat

libinput_seat_get_logical_name⚠
@ingroup seat

libinput_seat_get_physical_name⚠
@ingroup seat

libinput_seat_get_user_data⚠
@ingroup seat

libinput_seat_ref⚠
@ingroup seat

libinput_seat_set_user_data⚠
@ingroup seat

libinput_seat_unref⚠
@ingroup seat

libinput_set_user_data⚠
@ingroup base

libinput_suspend⚠
@ingroup base

libinput_tablet_pad_mode_group_button_is_toggle⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_get_index⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_get_mode⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_get_num_modes⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_get_user_data⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_has_button⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_has_ring⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_has_strip⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_ref⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_set_user_data⚠
@ingroup tablet_pad_modes

libinput_tablet_pad_mode_group_unref⚠
@ingroup tablet_pad_modes

libinput_tablet_tool_get_serial⚠
@ingroup event_tablet

libinput_tablet_tool_get_tool_id⚠
@ingroup event_tablet

libinput_tablet_tool_get_type⚠
@ingroup event_tablet

libinput_tablet_tool_get_user_data⚠
@ingroup event_tablet

libinput_tablet_tool_has_button⚠
@ingroup event_tablet

libinput_tablet_tool_has_distance⚠
@ingroup event_tablet

libinput_tablet_tool_has_pressure⚠
@ingroup event_tablet

libinput_tablet_tool_has_rotation⚠
@ingroup event_tablet

libinput_tablet_tool_has_slider⚠
@ingroup event_tablet

libinput_tablet_tool_has_tilt⚠
@ingroup event_tablet

libinput_tablet_tool_has_wheel⚠
@ingroup event_tablet

libinput_tablet_tool_is_unique⚠
@ingroup event_tablet

libinput_tablet_tool_ref⚠
@ingroup event_tablet

libinput_tablet_tool_set_user_data⚠
@ingroup event_tablet

libinput_tablet_tool_unref⚠
@ingroup event_tablet

libinput_udev_assign_seat⚠
@ingroup base

libinput_udev_create_context⚠
@ingroup base

libinput_unref⚠
@ingroup base

Type Definitions

__builtin_va_list

libinput_config_send_events_mode
@ingroup config

libinput_led
@ingroup device

libinput_log_handler
@ingroup base

va_list