Function input_sys::libinput_device_config_accel_set_speed

pub unsafe extern fn libinput_device_config_accel_set_speed(device: *mut libinput_device,
                                                            speed: f64)
                                                            -> libinput_config_status

@ingroup config

Set the pointer acceleration speed of this pointer device within a range of [-1, 1], where 0 is the default acceleration for this device, -1 is the slowest acceleration and 1 is the maximum acceleration available on this device. The actual pointer acceleration mechanism is implementation-dependent, as is the number of steps available within the range. libinput picks the semantically closest acceleration step if the requested value does not match a discrete setting.

@param device The device to configure @param speed The normalized speed, in a range of [-1, 1]

@return A config status code

@see libinput_device_config_accel_is_available @see libinput_device_config_accel_get_speed @see libinput_device_config_accel_get_default_speed