Function input_sys::libinput_device_config_left_handed_set

pub unsafe extern fn libinput_device_config_left_handed_set(device: *mut libinput_device,
                                                            left_handed: c_int)
                                                            -> libinput_config_status

@ingroup config

Set the left-handed configuration of the device.

The exact behavior is device-dependent. On a mouse and most pointing devices, left and right buttons are swapped but the middle button is unmodified. On a touchpad, physical buttons (if present) are swapped. On a clickpad, the top and bottom software-emulated buttons are swapped where present, the main area of the touchpad remains a left button. Tapping and clickfinger behavior is not affected by this setting.

Changing the left-handed configuration of a device may not take effect until all buttons have been logically released.

@param device The device to configure @param left_handed Zero to disable, non-zero to enable left-handed mode @return A configuration status code

@see libinput_device_config_left_handed_is_available @see libinput_device_config_left_handed_get @see libinput_device_config_left_handed_get_default