Function input_sys::libinput_device_config_scroll_has_natural_scroll

pub unsafe extern fn libinput_device_config_scroll_has_natural_scroll(device: *mut libinput_device)
                                                                      -> c_int

@ingroup config

Return non-zero if the device supports "natural scrolling".

In traditional scroll mode, the movement of fingers on a touchpad when scrolling matches the movement of the scroll bars. When the fingers move down, the scroll bar moves down, a line of text on the screen moves towards the upper end of the screen. This also matches scroll wheels on mice (wheel down, content moves up).

Natural scrolling is the term coined by Apple for inverted scrolling. In this mode, the effect of scrolling movement of fingers on a touchpad resemble physical manipulation of paper. When the fingers move down, a line of text on the screen moves down (scrollbars move up). This is the opposite of scroll wheels on mice.

A device supporting natural scrolling can be switched between traditional scroll mode and natural scroll mode.

@param device The device to configure

@return Zero if natural scrolling is not supported, non-zero if natural scrolling is supported by this device

@see libinput_device_config_scroll_set_natural_scroll_enabled @see libinput_device_config_scroll_get_natural_scroll_enabled @see libinput_device_config_scroll_get_default_natural_scroll_enabled