Function input_sys::libinput_device_config_tap_get_button_map

pub unsafe extern fn libinput_device_config_tap_get_button_map(device: *mut libinput_device)
                                                               -> libinput_config_tap_button_map

@ingroup config

Get the finger number to button number mapping for tap-to-click.

The return value for a device that does not support tapping is always @ref LIBINPUT_CONFIG_TAP_MAP_LRM.

@note It is an application bug to call this function for devices where libinput_device_config_tap_get_finger_count() returns 0.

@param device The device to configure @return The current finger-to-button number mapping

@see libinput_device_config_tap_set_button_map @see libinput_device_config_tap_get_default_button_map