Function input_sys::libinput_event_tablet_tool_get_tilt_x

pub unsafe extern fn libinput_event_tablet_tool_get_tilt_x(event: *mut libinput_event_tablet_tool)
                                                           -> f64

@ingroup event_tablet

Returns the current tilt along the X axis of the tablet's current logical orientation, in degrees off the tablet's z axis. That is, if the tool is perfectly orthogonal to the tablet, the tilt angle is 0. When the top tilts towards the logical top/left of the tablet, the x/y tilt angles are negative, if the top tilts towards the logical bottom/right of the tablet, the x/y tilt angles are positive.

If this axis does not exist on the current tool, this function returns 0.

@param event The libinput tablet tool event @return The current value of the axis in degrees