Function input_sys::libinput_event_touch_get_x

pub unsafe extern fn libinput_event_touch_get_x(event: *mut libinput_event_touch)
                                                -> f64

@ingroup event_touch

Return the current absolute x coordinate of the touch event, in mm from the top left corner of the device. To get the corresponding output screen coordinate, use libinput_event_touch_get_x_transformed().

For events not of type @ref LIBINPUT_EVENT_TOUCH_DOWN, @ref LIBINPUT_EVENT_TOUCH_MOTION, this function returns 0.

@note It is an application bug to call this function for events of type other than @ref LIBINPUT_EVENT_TOUCH_DOWN or @ref LIBINPUT_EVENT_TOUCH_MOTION.

@param event The libinput touch event @return The current absolute x coordinate