Function input_sys::libinput_path_create_context

pub unsafe extern fn libinput_path_create_context(interface: *const libinput_interface,
                                                  user_data: *mut c_void)
                                                  -> *mut libinput

@ingroup base

Create a new libinput context that requires the caller to manually add or remove devices with libinput_path_add_device() and libinput_path_remove_device().

The context is fully initialized but will not generate events until at least one device has been added.

The reference count of the context is initialized to 1. See @ref libinput_unref.

@param interface The callback interface @param user_data Caller-specific data passed to the various callback interfaces.

@return An initialized, empty libinput context.